Fix browser automation test budget
